+++ slash/utils/cli_utils.py
@@ -28,7 +28,9 @@ def add_pending_plugins_from_commandline(argv):
 
 def configure_arg_parser_by_plugins(parser):
     for plugin in itervalues(plugins.manager.get_installed_plugins()):
-        group = parser.add_argument_group('Options for --with-{0}'.format(plugin.get_name()))
+        group = parser.add_argument_group('Options for {}{}'.format(
+            '' if plugins.manager.is_internal_plugin(plugin) else '--with-',
+            plugin.get_name()))
         plugin.configure_argument_parser(group)
 
 def configure_arg_parser_by_config(parser, config=None):