+++ doc/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+* :release:`1.4.0 <8-8-2017>`
+* :feature:`-` Added new hook ``prepare_notification`` to process notifications before being sent by the notifications plugin
+* :feature:`662` Improve notifications plugin, add support for email notifications
+* :feature:`651` Add ``host_fqdn`` and ``host_name`` attributes to session
+* :feature:`647` Support internal plugins
+* :feature:`647` Support installing plugins as "internal" -- thus not letting users disable or enable them through the command line
 * :release:`1.3.0 <24-07-2017>`
 * :feature:`213` Added parallel execution capability (still considered experimental) - tests can be run in parallel by multiple subprocess "workers". See `the documentation <http://slash.readthedocs.io/en/master/parallel.html>`_ for more information
 * :feature:`596` Slash now supports a flag to disable assertion introspection on assertions containing messages (``run.message_assertion_introspection``)